+++ src/bp/ui-studio/src/web/translations/fr.json
@@ -145,6 +145,11 @@
         "add": "ajouter",
         "addAction": "Ajouter une nouvelle action",
         "cantSeeActions": "Vous ne voyez pas votre action?",
+        "chatbotExecutes": "L'assistant virtuel exécute",
+        "chatbotSays": "L'assistant virtuel dit",
+        "triggeredBy": "Le workflow est is déclenché par",
+        "workflowFails": "Échec du flux de travail",
+        "workflowSucceeds": "Réussite du flux de travail",
         "confirmOverwriteParameters": "Voulez-vous remplacer les paramètres existants?",
         "couldNotRetrieveActionServer": "Impossible de récupérer les serveurs d'actions",
         "editAction": "Modifier l'action",
