+++ src/bp/ui-studio/src/web/views/FlowBuilder/diagram/index.tsx
@@ -45,10 +45,10 @@ import storage from '~/util/storage'
 
 import { prepareEventForDiagram } from './debugger'
 import { defaultTransition, DiagramManager, DIAGRAM_PADDING, nodeTypes, Point } from './manager'
+import { BlockWidgetFactory } from './nodes/Block'
 import { DeletableLinkFactory } from './nodes/LinkWidget'
 import { SkillCallNodeModel, SkillCallWidgetFactory } from './nodes/SkillCallNode'
 import { StandardNodeModel, StandardWidgetFactory } from './nodes/StandardNode'
-import { BlockWidgetFactory } from './nodes_alb/Block'
 import style from './style.scss'
 import NodeToolbar from './NodeToolbar'
 import TriggerEditor from './TriggerEditor'
