+++ src/bp/ui-studio/src/web/views/FlowBuilder/diagram/nodes/Components/NodeContentItem.tsx
