+++ src/bp/ui-studio/src/web/views/FlowBuilder/diagram/nodes/SaySomethingContents/index.tsx
