+++ src/bp/ui-studio/src/web/views/FlowBuilder/diagram/nodes/SaySomethingContents/style.scss.d.ts
