+++ src/bp/ui-studio/src/web/views/FlowBuilder/diagram/nodes/TriggerContents/index.tsx
