--- src/bp/ui-studio/src/web/views/FlowBuilder/diagram/nodes_alb/StandardContents/style.scss
@@ -1,16 +1,0 @@
-.item {
-  word-wrap: break-word;
-  text-overflow: ellipsis;
-  max-height: 78px;
-  // overflow-y: scroll;
-  // border-bottom: 1px solid black;
-  font-size: 11px;
-}
-
-.section-next .item {
-  overflow: unset;
-}
-
-.item:last-child {
-  border-bottom: none;
-}
