--- src/bp/ui-studio/src/web/views/FlowBuilder/diagram/nodes_alb/StandardContents/style.scss.d.ts
@@ -1,8 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'item': string;
-  'section-next': string;
-}
-declare var cssExports: CssExports;
-export = cssExports;
